+++ app/assets/javascripts/decidim/decidim_awesome/awesome_admin.js
@@ -2,3 +2,4 @@
 // = require decidim/decidim_awesome/editors/quill_editor
 // = require decidim/decidim_awesome/editors/tabs_focus
 // = require decidim/decidim_awesome/admin/codemirror
+// = require decidim/decidim_awesome/admin/user_picker.js