+++ app/controllers/decidim/decidim_awesome/admin/config_controller.rb
@@ -11,7 +11,7 @@ module Decidim
 
         layout "decidim/admin/decidim_awesome"
 
-        helper_method :constraints_for
+        helper_method :constraints_for, :users_for
         before_action do
           enforce_permission_to :edit_config, configs
         end
@@ -35,32 +35,19 @@ module Decidim
           end
         end
 
-        def new_scoped_style
-          CreateScopedStyle.call(current_organization) do
-            on(:ok) do |key|
-              flash[:notice] = I18n.t("config.create_scoped_style.success", key: key, scope: "decidim.decidim_awesome.admin")
-            end
+        def users
+          respond_to do |format|
+            format.json do
+              if (term = params[:term].to_s).present?
+                query = current_organization.users.order(name: :asc)
+                query = query.where("name ILIKE :term OR nickname ILIKE :term OR email ILIKE :term", term: "%#{term}%")
 
-            on(:invalid) do |message|
-              flash[:alert] = I18n.t("config.create_scoped_style.error", error: message, scope: "decidim.decidim_awesome.admin")
-            end
-          end
-
-          redirect_to decidim_admin_decidim_awesome.config_path(:styles)
-        end
-
-        def destroy_scoped_style
-          DestroyScopedStyle.call(params[:key], current_organization) do
-            on(:ok) do |key|
-              flash[:notice] = I18n.t("config.destroy_scoped_style.success", key: key, scope: "decidim.decidim_awesome.admin")
-            end
-
-            on(:invalid) do |message|
-              flash[:alert] = I18n.t("config.destroy_scoped_style.error", error: message, scope: "decidim.decidim_awesome.admin")
+                render json: query.all.collect { |u| { id: u.id, text: format_user_name(u) } }
+              else
+                render json: []
+              end
             end
           end
-
-          redirect_to decidim_admin_decidim_awesome.config_path(:styles)
         end
 
         private
@@ -73,6 +60,14 @@ module Decidim
           return params[:config].keys if params.has_key?(:config)
 
           DecidimAwesome.config.keys
+        end
+
+        def users_for(ids_list)
+          Decidim::User.where(id: ids_list).map { |user| OpenStruct.new(text: format_user_name(user), id: user.id) }
+        end
+
+        def format_user_name(user)
+          "<span class='#{"is-admin" if user.read_attribute("admin")}'>#{user.name} (@#{user.nickname} - #{user.email})</span>"
         end
       end
     end