+++ app/permissions/decidim/decidim_awesome/admin/permissions.rb
@@ -8,6 +8,8 @@ module Decidim
 
         def permissions
           return permission_action if permission_action.scope != :admin
+          return permission_action unless user
+          return permission_action if user.read_attribute("admin").blank?
 
           toggle_allow(config_enabled?(permission_action.subject)) if permission_action.action == :edit_config